Resolves #55 revamp backend structure
