+++ backend/package.json
@@ -19,8 +19,12 @@
     "xlsx": "^0.12.7"
   },
   "devDependencies": {
+    "chai": "^4.1.2",
     "lint-staged": "^7.2.0",
+    "mocha": "^5.2.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^1.13.5"
+    "prettier": "^1.13.5",
+    "sinon": "^6.0.1",
+    "supertest": "^3.1.0"
   }
 }
