+++ backend/src/api/index.js
@@ -1,0 +1,9 @@
+const candidates = require('./candidates')
+const matchCandidates = require('./matchCandidates')
+const matches = require('./matches')
+
+module.exports = {
+  candidates,
+  matchCandidates,
+  matches
+}
